Clean up weird title handling and output
I'm not sure why this was doing these weird things to begin with. Anyway,
this code is mostly the same that's already present in MonoBook/Vector/any
sane, modern custom skin, so this should be safe. At least I didn't notice
any changes in the visual output while testing this out.