Rename misleadingly named local variable
The wg prefix stands for "wiki global" and as such it should only be used
for local variables. This array, now called
$defaultPurpleFeatureConfiguration, populates $wgPurpleFeatures with some
default values unless there is already a different override in place in
LocalSettings.php, but this variable itself isn't a global per se, so it
shouldn't really have the wg prefix.